pyproject.toml
[tool.poetry]
name = "code-names-bot-text-processor"
version = "0.1.0"
description = ""
authors = ["Nalu Zou <zonalu@amazon.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
spacy = "^3.4.1"
nltk = "^3.7"
black = "^22.6.0"
isort = "^5.10.1"
PyYAML = "^6.0"
PySide6 = "^6.3.1"
scikit-learn = "^1.1.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
format                  = "tasks:format"
setup                   = "tasks:setup"

dependency-parse        = "tools.dependency_parse:main"
show-named-entities     = "tools.show_named_entities:main"
custom-parse            = "tools.custom_parse:main"
parse-key-terms         = "tools.parse_key_terms:main"
get-oxford-definition   = "tools.get_oxford_definition:main"
list-lemmas             = "tools.list_lemmas:main"
test-term-chunker       = "tools.test_term_chunker:main"
get-synsets             = "tools.get_synsets:main"
download-terms          = "tools.download_terms:main"
process-labels          = "tools.process_labels:main"

labeler                 = "labeler.labeler:main"
view-predictions        = "labeler.predicted_labels_viewer:main"

load-definitions        = "data_scripts.load_definitions:main"
process-compounds       = "data_scripts.compounds_processor:main"
calc-doc-freq           = "data_scripts.calc_doc_frequencies:main"

test-single             = "evaluation.test_single:main"
evaluate                = "evaluation.evaluate:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
